+++ src/collective/contact/plonegroup/browser/settings.py
@@ -337,16 +337,22 @@ def selectedOrganizationsPloneGroupsVocabulary(functions=[], group_title=True):
     return SimpleVocabulary(terms)
 
 
+def getSelectedOrganizations(separator=' - ', first_index=1):
+    """ Return a list of tuples (uid, title) """
+    ret = []
+    registry = getUtility(IRegistry)
+    for uid in registry[ORGANIZATIONS_REGISTRY]:
+        title = uuidToObject(uid).get_full_title(separator=separator, first_index=first_index)
+        ret.append((uid, title))
+    return ret
+
+
 @ram.cache(lambda *args: True)  # not used
 def selectedOrganizationsVocabulary():
     """
         Returns a vocabulary of selected organizations
     """
-    registry = getUtility(IRegistry)
-    terms = []
-    for uid in registry[ORGANIZATIONS_REGISTRY]:
-        title = uuidToObject(uid).get_full_title(separator=' - ', first_index=1)
-        terms.append(SimpleTerm(uid, token=uid, title=title))
+    terms = [SimpleTerm(t[0], title=t[1]) for t in getSelectedOrganizations()]
     return SimpleVocabulary(terms)